pyproject.toml
[project]
name = "crypto-us-yields"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dune-client>=1.10.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "streamlit>=1.33.0",
    "watchdog>=6.0.0",
]
